#  Change Chunk.Metric from a map to a slice, for performance