--- a/pkg/chunk/json_helpers.go
+++ b/pkg/chunk/json_helpers.go
@@ -5,24 +5,45 @@ import (
 
 	jsoniter "github.com/json-iterator/go"
 	"github.com/prometheus/common/model"
+	"github.com/prometheus/prometheus/pkg/labels"
 )
 
 func init() {
-	jsoniter.RegisterTypeDecoderFunc("model.Metric", decodeMetric)
+	jsoniter.RegisterTypeDecoderFunc("labels.Labels", decodeLabels)
+	jsoniter.RegisterTypeEncoderFunc("labels.Labels", encodeLabels, labelsIsEmpty)
 	jsoniter.RegisterTypeDecoderFunc("model.Time", decodeModelTime)
 	jsoniter.RegisterTypeEncoderFunc("model.Time", encodeModelTime, modelTimeIsEmpty)
 }
 
-// decoding model.Metric via ReadMapCB is faster than the generic jsoniter
-// decoder because the latter allocates memory for each string via reflect.
-func decodeMetric(ptr unsafe.Pointer, iter *jsoniter.Iterator) {
-	mapPtr := (*model.Metric)(ptr)
-	*mapPtr = make(model.Metric, 10)
+// Override Prometheus' labels.Labels decoder which goes via a map
+func decodeLabels(ptr unsafe.Pointer, iter *jsoniter.Iterator) {
+	labelsPtr := (*labels.Labels)(ptr)
+	*labelsPtr = make(labels.Labels, 0, 10)
 	iter.ReadMapCB(func(iter *jsoniter.Iterator, key string) bool {
 		value := iter.ReadString()
-		(*mapPtr)[model.LabelName(key)] = model.LabelValue(value)
+		*labelsPtr = append(*labelsPtr, labels.Label{Name: key, Value: value})
 		return true
 	})
+}
+
+// Override Prometheus' labels.Labels encoder which goes via a map
+func encodeLabels(ptr unsafe.Pointer, stream *jsoniter.Stream) {
+	labelsPtr := (*labels.Labels)(ptr)
+	stream.WriteObjectStart()
+	for i, v := range *labelsPtr {
+		if i != 0 {
+			stream.WriteMore()
+		}
+		stream.WriteString(v.Name)
+		stream.WriteRaw(`:`)
+		stream.WriteString(v.Value)
+	}
+	stream.WriteObjectEnd()
+}
+
+func labelsIsEmpty(ptr unsafe.Pointer) bool {
+	labelsPtr := (*labels.Labels)(ptr)
+	return len(*labelsPtr) == 0
 }
 
 // Decode via jsoniter's float64 routine is faster than getting the string data and decoding as two integers