--- a/pkg/ingester/client/fnv.go
+++ b/pkg/ingester/client/fnv.go
@@ -38,6 +38,15 @@ func hashAdd(h uint64, s string) uint64 {
 	return h
 }
 
+// hashAdd adds a string to a fnv64a hash value, returning the updated hash.
+func hashAddString(h uint64, s string) uint64 {
+	for i := 0; i < len(s); i++ {
+		h ^= uint64(s[i])
+		h *= prime64
+	}
+	return h
+}
+
 // hashAddByte adds a byte to a fnv64a hash value, returning the updated hash.
 func hashAddByte(h uint64, b byte) uint64 {
 	h ^= uint64(b)