This continues what was started in #1007, so now we never create a `map` for labels in the ingester.

I started on this when I noticed that index caching was taking a lot of time in `Metric.String()`.  This PR reduces that a bit, but there are further improvements that can be made to `Labels.String()`.

While fixing up the tests I noticed we go through all kinds of indirections, from chunks to matrix to series set and back again, and I started to remove some of those.  That's a tough job though, so I stopped after a few.

Benchmarks (best of 3):
Before:
```
BenchmarkEncode      	  300000	      4137 ns/op	    2978 B/op	      10 allocs/op
BenchmarkDecode1     	   50000	     29984 ns/op	  152483 B/op	      57 allocs/op
BenchmarkDecode100   	    2000	    692953 ns/op	  619859 B/op	    5234 allocs/op
BenchmarkDecode10000 	      20	 102851651 ns/op	49305985 B/op	  543190 allocs/op
BenchmarkIndexCaching 	   30000	     60016 ns/op	   12816 B/op	     188 allocs/op
```

After:
```
BenchmarkEncode      	  500000	      3699 ns/op	    2850 B/op	       8 allocs/op
BenchmarkDecode1     	   50000	     24661 ns/op	  151576 B/op	      56 allocs/op
BenchmarkDecode100   	    2000	    543184 ns/op	  548370 B/op	    5303 allocs/op
BenchmarkDecode10000 	      20	  73501908 ns/op	40227785 B/op	  530003 allocs/op
BenchmarkIndexCaching 	   30000	     43642 ns/op	   14209 B/op	     128 allocs/op
```