@tomwilkie 2019-05-09
Could we have a test to confirm that labels.Labels can be unmarsalled from a JSON map?  Something in chunk_test.go that unmarshals a string into a chunk, to show we haven't broken compatibility?