@bboreham 2019-05-09
> unmarshals a string into a chunk

a hard-coded string generated from the previous code?