@tomwilkie 2019-05-09
> a hard-coded string generated from the previous code?

Yeah, I can't think of a better way!  Or download a chunk and stick it in the repo as a file.