@bboreham 2019-05-12
I have addressed review comments, added a backwards-compatibility test and fixed the bug that it revealed.